view and create hotkeys from trade page
