+++ components/trade/AdvancedMarketHeader.tsx
@@ -15,6 +15,7 @@ import SpotMarketDetailsModal from '@components/modals/SpotMarketDetailsModal'
 import { MarketData } from 'types'
 import MarketChange from '@components/shared/MarketChange'
 import useMarketsData from 'hooks/useMarketsData'
+import HotKeysDrawerButton from './HotKeysDrawerButton'
 
 const AdvancedMarketHeader = ({
   showChart,
@@ -136,9 +137,10 @@ const AdvancedMarketHeader = ({
               </div>
             )}
           </div>
-          <div className="ml-6 flex items-center space-x-4">
+          <div className="ml-6 flex items-center">
+            <HotKeysDrawerButton />
             <LinkButton
-              className="flex items-center whitespace-nowrap text-th-fgd-3"
+              className="ml-4 flex items-center whitespace-nowrap text-th-fgd-3"
               onClick={() => setShowMarketDetails(true)}
             >
               <InformationCircleIcon className="h-5 w-5 flex-shrink-0 md:mr-1.5 md:h-4 md:w-4" />
@@ -148,7 +150,9 @@ const AdvancedMarketHeader = ({
             </LinkButton>
             {setShowChart ? (
               <IconButton
-                className={showChart ? 'text-th-active' : 'text-th-fgd-2'}
+                className={`ml-4 ${
+                  showChart ? 'text-th-active' : 'text-th-fgd-2'
+                }`}
                 onClick={() => setShowChart(!showChart)}
                 hideBg
               >
