+++ public/locales/ru/settings.json
@@ -37,7 +37,7 @@
     "english": "English",
     "high-contrast": "High Contrast",
     "hot-keys": "Hot Keys",
-    "hot-keys-desc": "Use hot keys to place orders on the trade page. They execute on the market you're viewing and are not market specific.",
+    "hot-keys-desc": "Use hot keys to place orders quickly. They execute on the market you're viewing and are not market specific.",
     "increase-account-size": "Increase Account Size",
     "key-sequence": "Key Sequence",
     "language": "Language",
